Move test scripts to another folder
