+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/1888_1_navi.lua
@@ -2,14 +2,14 @@
 -- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1888
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/Policies_Security/Trigger_PTU_NO_Certificate/common')
+local common = require('test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Variables ]]
-local serviceId = 11
+local serviceId = 7
 local appHMIType = "NAVIGATION"
 
 --[[ General configuration parameters ]]
@@ -37,6 +37,23 @@ local function startServiceSecured(pData)
   common.delayedExp()
 end
 
+local function sendRPCAddCommandSecured()
+  local params = {
+    cmdID = 1,
+    menuParams = {
+      position = 1,
+      menuName = "Command_1"
+    }
+  }
+  local cid = common.getMobileSession():SendEncryptedRPC("AddCommand", params)
+  common.getHMIConnection():ExpectRequest("UI.AddCommand", params)
+  :Do(function(_, data)
+      common.getHMIConnection():SendResponse(data.id, data.method, "SUCCESS", { })
+    end)
+  common.getMobileSession():ExpectEncryptedResponse(cid, { success = true, resultCode = "SUCCESS" })
+  common.getMobileSession():ExpectEncryptedNotification("OnHashChange")
+end
+
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
@@ -48,11 +65,17 @@ runner.Title("Test")
 runner.Step("Register App", common.registerApp)
 runner.Step("Activate App", common.activateApp)
 
+runner.Step("StartService Secured NACK", startServiceSecured, { {
+  frameInfo = common.frameInfo.START_SERVICE_NACK,
+  encryption = false } })
+
 runner.Step("PolicyTableUpdate with certificate", common.policyTableUpdate, { ptUpdate })
 
 runner.Step("StartService Secured ACK", startServiceSecured, { {
   frameInfo = common.frameInfo.START_SERVICE_ACK,
   encryption = true } })
+
+runner.Step("AddCommand Secured", sendRPCAddCommandSecured)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
