+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/1888_5_navi_audio_force_on.lua
@@ -1,55 +1,58 @@
 ---------------------------------------------------------------------------------------------------
--- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1891
+-- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1888
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/Policies_Security/Trigger_PTU_NO_Certificate/common')
+local common = require('test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Variables ]]
-local serviceId = 11
+local serviceId = 10
 local appHMIType = "NAVIGATION"
 
 --[[ General configuration parameters ]]
+config.application1.registerAppInterfaceParams.appName = "server"
+config.application1.registerAppInterfaceParams.appID = "SPT"
 config.application1.registerAppInterfaceParams.appHMIType = { appHMIType }
 
 --[[ Local Functions ]]
 local function ptUpdate(pTbl)
+  local filePath = "./files/Security/client_credential.pem"
+  local crt = common.readFile(filePath)
+  pTbl.policy_table.module_config.certificate = crt
   pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { appHMIType }
-  pTbl.policy_table.module_config.seconds_between_retries = nil
 end
 
-local function expNotificationFunc()
-  common.getHMIConnection():ExpectNotification("SDL.OnStatusUpdate",
-    { status = "UPDATE_NEEDED" }, { status = "UPDATING" })
-  :Times(2)
-end
-
-local function startServiceSecured()
+local function startServiceSecured(pData)
   common.getMobileSession():StartSecureService(serviceId)
-  common.getMobileSession():ExpectControlMessage(serviceId, {
-    frameInfo = common.frameInfo.START_SERVICE_ACK,
-    encryption = false
-  })
+  common.getMobileSession():ExpectControlMessage(serviceId, pData)
 
+  local handshakeOccurences = 0
+  if pData.encryption == true then handshakeOccurences = 1 end
   common.getMobileSession():ExpectHandshakeMessage()
-  :Times(0)
+  :Times(handshakeOccurences)
+
   common.delayedExp()
 end
 
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
-runner.Step("Set ForceProtectedService OFF", common.setForceProtectedServiceParam, { "Non" })
+runner.Step("Set ForceProtectedService ON", common.setForceProtectedServiceParam, { "0x0A" })
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 
 runner.Title("Test")
+
 runner.Step("Register App", common.registerApp)
-runner.Step("PolicyTableUpdate fails", common.policyTableUpdate, { ptUpdate, expNotificationFunc })
 runner.Step("Activate App", common.activateApp)
-runner.Step("StartService Secured ACK, no encryption, no Handshake", startServiceSecured)
+
+runner.Step("PolicyTableUpdate with certificate", common.policyTableUpdate, { ptUpdate })
+
+runner.Step("StartService Secured ACK", startServiceSecured, { {
+  frameInfo = common.frameInfo.START_SERVICE_ACK,
+  encryption = true } })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
