+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/1894_4_navi_video.lua
@@ -1,48 +1,34 @@
 ---------------------------------------------------------------------------------------------------
--- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1891
+-- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1894
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/Policies_Security/Trigger_PTU_NO_Certificate/common')
+local common = require('test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Variables ]]
-local serviceId = 7
-local appHMIType = "DEFAULT"
+local serviceId = 11
+local appHMIType = "NAVIGATION"
 
 --[[ General configuration parameters ]]
 config.application1.registerAppInterfaceParams.appHMIType = { appHMIType }
 
 --[[ Local Functions ]]
-local function ptUpdateSuccess(pTbl)
-  pTbl.policy_table.module_config.certificate = nil
+local function ptUpdate(pTbl)
+	pTbl.policy_table.module_config.certificate = nil
   pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { appHMIType }
-end
-
-local function ptUpdateUnssucess(pTbl)
-  pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { appHMIType }
-  pTbl.policy_table.module_config.seconds_between_retries = nil
 end
 
 local function startServiceSecured()
   common.getMobileSession():StartSecureService(serviceId)
   common.getMobileSession():ExpectControlMessage(serviceId, {
-    frameInfo = common.frameInfo.START_SERVICE_NACK,
+    frameInfo = common.frameInfo.START_SERVICE_ACK,
     encryption = false
   })
   common.getMobileSession():ExpectHandshakeMessage()
   :Times(0)
-
-  local function expNotificationFunc()
-    common.getHMIConnection():ExpectNotification("SDL.OnStatusUpdate",
-      { status = "UPDATE_NEEDED" }, { status = "UPDATING" },
-      { status = "UPDATE_NEEDED" }, { status = "UPDATING" })
-    :Times(4)
-  end
-
-  common.policyTableUpdate(ptUpdateUnssucess, expNotificationFunc)
   common.delayedExp()
 end
 
@@ -54,8 +40,9 @@ runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 
 runner.Title("Test")
 runner.Step("Register App", common.registerApp)
-runner.Step("PolicyTableUpdate fails", common.policyTableUpdate, { ptUpdateSuccess })
-runner.Step("StartService Secured, PTU started and fails, NACK, no Handshake", startServiceSecured)
+runner.Step("PolicyTableUpdate without certificate", common.policyTableUpdate, { ptUpdate })
+runner.Step("Activate App", common.activateApp)
+runner.Step("StartService Secured ACK, no encryption, no Handshake", startServiceSecured)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
