+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/1922_2_navi_video.lua
@@ -1,15 +1,15 @@
 ---------------------------------------------------------------------------------------------------
--- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1924
+-- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1922
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/Policies_Security/Trigger_PTU_NO_Certificate/common')
+local common = require('test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Variables ]]
-local serviceId = 7
+local serviceId = 11
 local appHMIType = "NAVIGATION"
 
 --[[ General configuration parameters ]]
@@ -27,26 +27,22 @@ local function startServiceSecured()
     frameInfo = common.frameInfo.START_SERVICE_NACK,
     encryption = false
   })
-
-  common.getHMIConnection():ExpectNotification("SDL.OnStatusUpdate",
-    { status = "UPDATE_NEEDED" }, { status = "UPDATING" })
+  common.getMobileSession():ExpectHandshakeMessage()
   :Times(0)
-  common.getHMIConnection():ExpectRequest("BasicCommunication.PolicyUpdate")
-  :Times(0)
-
   common.delayedExp()
 end
 
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
+runner.Step("Set ForceProtectedService ON", common.setForceProtectedServiceParam, { "0x0B" })
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
-runner.Step("Register App", common.registerApp)
-runner.Step("PolicyTableUpdate without certificate", common.policyTableUpdate, { ptUpdate })
 
 runner.Title("Test")
-
-runner.Step("StartService Secured, PTU not started, NACK", startServiceSecured)
+runner.Step("Register App", common.registerApp)
+runner.Step("PolicyTableUpdate without certificate", common.policyTableUpdate, { ptUpdate })
+runner.Step("Activate App", common.activateApp)
+runner.Step("StartService Secured NACK, no Handshake", startServiceSecured)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
