+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/1923_1_navi_audio.lua
@@ -1,58 +1,54 @@
 ---------------------------------------------------------------------------------------------------
--- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1888
+-- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1923
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/Policies_Security/Trigger_PTU_NO_Certificate/common')
+local common = require('test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Variables ]]
-local serviceId = 11
+local serviceId = 10
 local appHMIType = "NAVIGATION"
 
 --[[ General configuration parameters ]]
-config.application1.registerAppInterfaceParams.appName = "server"
-config.application1.registerAppInterfaceParams.appID = "SPT"
 config.application1.registerAppInterfaceParams.appHMIType = { appHMIType }
 
 --[[ Local Functions ]]
 local function ptUpdate(pTbl)
-  local filePath = "./files/Security/client_credential.pem"
-  local crt = common.readFile(filePath)
-  pTbl.policy_table.module_config.certificate = crt
   pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { appHMIType }
+	pTbl.policy_table.module_config.seconds_between_retries = nil
 end
 
-local function startServiceSecured(pData)
+local function startServiceSecured()
   common.getMobileSession():StartSecureService(serviceId)
-  common.getMobileSession():ExpectControlMessage(serviceId, pData)
-
-  local handshakeOccurences = 0
-  if pData.encryption == true then handshakeOccurences = 1 end
+  common.getMobileSession():ExpectControlMessage(serviceId, {
+    frameInfo = common.frameInfo.START_SERVICE_NACK,
+    encryption = false
+  })
   common.getMobileSession():ExpectHandshakeMessage()
-  :Times(handshakeOccurences)
-
+  :Times(0)
   common.delayedExp()
+end
+
+local function expNotificationFunc()
+  common.getHMIConnection():ExpectNotification("SDL.OnStatusUpdate",
+    { status = "UPDATE_NEEDED" }, { status = "UPDATING" })
+  :Times(2)
 end
 
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
-runner.Step("Set ForceProtectedService ON", common.setForceProtectedServiceParam, { "0x0B" })
+runner.Step("Set ForceProtectedService ON", common.setForceProtectedServiceParam, { "0x0A" })
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 
 runner.Title("Test")
-
 runner.Step("Register App", common.registerApp)
+runner.Step("PolicyTableUpdate fails", common.policyTableUpdate, { ptUpdate, expNotificationFunc })
 runner.Step("Activate App", common.activateApp)
-
-runner.Step("PolicyTableUpdate with certificate", common.policyTableUpdate, { ptUpdate })
-
-runner.Step("StartService Secured ACK", startServiceSecured, { {
-  frameInfo = common.frameInfo.START_SERVICE_ACK,
-  encryption = true } })
+runner.Step("StartService Secured NACK, no Handshake", startServiceSecured)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
