+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/1924_1_non-navi.lua
@@ -1,55 +1,49 @@
 ---------------------------------------------------------------------------------------------------
--- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1891
+-- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1924
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/Policies_Security/Trigger_PTU_NO_Certificate/common')
+local common = require('test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Variables ]]
-local serviceId = 10
-local appHMIType = "NAVIGATION"
+local serviceId = 7
+local appHMIType = "DEFAULT"
 
 --[[ General configuration parameters ]]
 config.application1.registerAppInterfaceParams.appHMIType = { appHMIType }
 
 --[[ Local Functions ]]
 local function ptUpdate(pTbl)
+	pTbl.policy_table.module_config.certificate = nil
   pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { appHMIType }
-  pTbl.policy_table.module_config.seconds_between_retries = nil
-end
-
-local function expNotificationFunc()
-  common.getHMIConnection():ExpectNotification("SDL.OnStatusUpdate",
-    { status = "UPDATE_NEEDED" }, { status = "UPDATING" })
-  :Times(2)
 end
 
 local function startServiceSecured()
   common.getMobileSession():StartSecureService(serviceId)
-  common.getMobileSession():ExpectControlMessage(serviceId, {
-    frameInfo = common.frameInfo.START_SERVICE_ACK,
-    encryption = false
-  })
-
-  common.getMobileSession():ExpectHandshakeMessage()
+  common.getMobileSession():ExpectControlMessage(serviceId, { })
   :Times(0)
+
+  common.getHMIConnection():ExpectNotification("SDL.OnStatusUpdate",
+    { status = "UPDATE_NEEDED" }, { status = "UPDATING" })
+  :Times(2)
+  common.getHMIConnection():ExpectRequest("BasicCommunication.PolicyUpdate")
+
   common.delayedExp()
 end
 
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
-runner.Step("Set ForceProtectedService OFF", common.setForceProtectedServiceParam, { "Non" })
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
+runner.Step("Register App", common.registerApp)
+runner.Step("PolicyTableUpdate without certificate", common.policyTableUpdate, { ptUpdate })
 
 runner.Title("Test")
-runner.Step("Register App", common.registerApp)
-runner.Step("PolicyTableUpdate fails", common.policyTableUpdate, { ptUpdate, expNotificationFunc })
-runner.Step("Activate App", common.activateApp)
-runner.Step("StartService Secured ACK, no encryption, no Handshake", startServiceSecured)
+
+runner.Step("StartService Secured, PTU started, No ACK/NACK", startServiceSecured)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
