+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/1924_2_navi.lua
@@ -2,7 +2,7 @@
 -- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1924
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/Policies_Security/Trigger_PTU_NO_Certificate/common')
+local common = require('test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
@@ -10,26 +10,29 @@ runner.testSettings.isSelfIncluded = false
 
 --[[ Local Variables ]]
 local serviceId = 7
-local appHMIType = "DEFAULT"
+local appHMIType = "NAVIGATION"
 
 --[[ General configuration parameters ]]
 config.application1.registerAppInterfaceParams.appHMIType = { appHMIType }
 
 --[[ Local Functions ]]
 local function ptUpdate(pTbl)
-	pTbl.policy_table.module_config.certificate = nil
+  pTbl.policy_table.module_config.certificate = nil
   pTbl.policy_table.app_policies[common.getAppID()].AppHMIType = { appHMIType }
 end
 
 local function startServiceSecured()
   common.getMobileSession():StartSecureService(serviceId)
-  common.getMobileSession():ExpectControlMessage(serviceId, { })
-  :Times(0)
+  common.getMobileSession():ExpectControlMessage(serviceId, {
+    frameInfo = common.frameInfo.START_SERVICE_NACK,
+    encryption = false
+  })
 
   common.getHMIConnection():ExpectNotification("SDL.OnStatusUpdate",
     { status = "UPDATE_NEEDED" }, { status = "UPDATING" })
-  :Times(2)
+  :Times(0)
   common.getHMIConnection():ExpectRequest("BasicCommunication.PolicyUpdate")
+  :Times(0)
 
   common.delayedExp()
 end
@@ -43,7 +46,7 @@ runner.Step("PolicyTableUpdate without certificate", common.policyTableUpdate, {
 
 runner.Title("Test")
 
-runner.Step("StartService Secured, PTU started, No ACK/NACK", startServiceSecured)
+runner.Step("StartService Secured, PTU not started, NACK", startServiceSecured)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
