+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/1925_8_non-navi_cert_existing_app.lua
@@ -2,7 +2,7 @@
 -- Issue: https://github.com/SmartDeviceLink/sdl_core/issues/1925
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
-local common = require('test_scripts/Policies/Policies_Security/Trigger_PTU_NO_Certificate/common')
+local common = require('test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common')
 local runner = require('user_modules/script_runner')
 
 --[[ Test Configuration ]]
@@ -11,7 +11,7 @@ runner.testSettings.isSelfIncluded = false
 --[[ Local Variables ]]
 local appHMIType = {
   [1] = "DEFAULT",
-  [2] = "NAVIGATION"
+  [2] = "DEFAULT"
 }
 
 --[[ General configuration parameters ]]
@@ -20,8 +20,10 @@ config.application2.registerAppInterfaceParams.appHMIType = { appHMIType[2] }
 
 --[[ Local Functions ]]
 local function ptUpdate(pTbl)
-  pTbl.policy_table.module_config.certificate = nil
   pTbl.policy_table.app_policies[common.getAppID(1)].AppHMIType = { appHMIType[1] }
+  local filePath = "./files/Security/client_credential.pem"
+  local crt = common.readFile(filePath)
+  pTbl.policy_table.module_config.certificate = crt
   common.updatePTU(pTbl, 2)
   pTbl.policy_table.app_policies[common.getAppID(2)].AppHMIType = { appHMIType[2] }
 end
@@ -29,9 +31,11 @@ end
 local function registerApp(pAppId)
   common.getHMIConnection():ExpectNotification("SDL.OnStatusUpdate",
     { status = "UPDATE_NEEDED" }, { status = "UPDATING" })
-  :Times(2)
+  :Times(0)
   common.getHMIConnection():ExpectRequest("BasicCommunication.PolicyUpdate")
+  :Times(0)
   common.registerAppWOPTU(pAppId)
+  common.delayedExp()
 end
 
 --[[ Scenario ]]
@@ -42,7 +46,7 @@ runner.Step("Register " .. appHMIType[1] .. " App", common.registerApp, { 1 })
 runner.Step("PTU 1 finished", common.policyTableUpdate, { ptUpdate })
 
 runner.Title("Test")
-runner.Step("Register " .. appHMIType[2] .. " App, PTU started", registerApp, { 2 })
+runner.Step("Register " .. appHMIType[2] .. " App, PTU not started", registerApp, { 2 })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
