+++ test_scripts/Defects/4_5/Trigger_PTU_NO_Certificate/common.lua
